+++ exclude.c
@@ -1665,24 +1665,7 @@ void parse_filter_file(filter_rule_list *listp, const char *fname, const filter_
 			open_path = line;
 		} else
 			open_path = fname;
-
-		/* Confine the open to the module root.  The ownership walk on its own
-		 * is not enough for a peer-driven merge file: a non-chrooted daemon
-		 * writes --backup-dir entries as root, so a raced backup symlink is
-		 * ROOT-owned -- exactly what open_no_attacker_symlinks() treats as
-		 * trusted -- and naming it in a dir-merge rule would read an
-		 * out-of-module file in as filter rules (their text comes back to the
-		 * peer in "Unknown filter rule" errors).
-		 *
-		 * The daemon's own "filter"/"include from"/"exclude from" parameters
-		 * are exempt: those are operator-configured and legitimately live
-		 * outside the module (/etc/rsync/excludes and the like). */
-		int save_opr = operator_path_resolve;
-		if (!daemon_config_filter_file)
-			operator_path_resolve = 1;
-		fd = open_no_attacker_symlinks(open_path, O_RDONLY, 0);
-		operator_path_resolve = save_opr;
-
+		fd = vfs_open_owner_walk(open_path, O_RDONLY, 0);
 		if (fd < 0)
 			fp = NULL;
 		else if (!(fp = fdopen(fd, "rb")))
