+++ syscall.c
@@ -53,372 +53,11 @@ extern int preserve_executability;
 extern int open_noatime;
 extern int copy_links;
 extern int copy_unsafe_links;
-extern int am_daemon;
 extern int insecure_links;
 extern int module_id;
 
 
 
-
-/* Advance the tracked absolute path `abspath` by one resolved component,
- * normalizing "." and ".." exactly as openat() does so the module-confinement
- * check (abspath_outside_confinement) sees the REAL resolved target.  -1/
- * ENAMETOOLONG on overflow. */
-static int abspath_step(char *abspath, size_t cap, const char *comp, size_t comp_len)
-{
-	if (comp_len == 1 && comp[0] == '.')
-		return 0;				/* "." -- no movement */
-	if (comp_len == 2 && comp[0] == '.' && comp[1] == '.') {
-		char *s = strrchr(abspath, '/');	/* ".." -- pop a component */
-		if (s)
-			*s = '\0';
-		else
-			abspath[0] = '\0';
-		return 0;
-	}
-	size_t al = strlen(abspath);
-	size_t off = (al > 0 && abspath[al-1] == '/') ? al : al + 1;	/* no "//" */
-	if (off + comp_len >= cap) {
-		errno = ENAMETOOLONG;
-		return -1;
-	}
-	if (off != al)
-		abspath[al] = '/';
-	memcpy(abspath + off, comp, comp_len + 1);
-	return 0;
-}
-
-/* Open an operator-supplied path, refusing to traverse any symlink (parent or
- * leaf) not owned by uid 0 or our euid.  A trusted-owned symlink (e.g. root's
- * /var/log -> /data/log) is still followed; an untrusted one fails ELOOP.
- * Unlike plain O_NOFOLLOW this also defends a planted parent component
- * (--log-file=$plant/log), not just a planted leaf.  Used for opens that may
- * transit attacker-writable parents: --log-file, --password-file, --*-from,
- * --read/write-batch, daemon motd/lock/early-input/--config.
- *
- * Walks component-by-component with fstatat(AT_SYMLINK_NOFOLLOW) +
- * openat(O_NOFOLLOW), splicing a trusted symlink's target back into the path.
- * Returns the fd, or -1 (errno ELOOP on the security refusal so callers can
- * tell it apart).  Falls back to plain open() where openat/O_NOFOLLOW are
- * unavailable. */
-/* Core walk.  When out_abs is non-NULL and the path resolves to a directory
- * (O_DIRECTORY), the resolved absolute path is copied there -- owner_walk_parent
- * uses it to filter-check the (otherwise unchecked) leaf basename. */
-static int ona_open(const char *path, int flags, mode_t mode, char *out_abs, size_t out_cap)
-{
-#if defined AT_FDCWD && defined O_NOFOLLOW
-	/* O_CLOEXEC predates some still-supported targets; mirror rand_bytes()'s
-	 * fallback in syscall.c so a build without it still compiles. */
-#ifndef O_CLOEXEC
-#define O_CLOEXEC 0
-#endif
-	if (!path || !*path) {
-		errno = EINVAL;
-		return -1;
-	}
-
-	/* Opted out (local --insecure-links, or a daemon module with "insecure
-	 * links = yes"): restore the legacy symlink-following open. */
-	if (vfs_symlink_optout_allowed())
-		return open(path, flags, mode);
-
-	const uid_t trusted_uid = geteuid();
-	int dfd = AT_FDCWD;
-	int dfd_owns = 0;
-
-	/* Absolute path of the current dir, for the confinement refusal
-	 * (abspath_outside_confinement).  A relative operator path starts at the
-	 * daemon's cwd == the module root; an absolute one (or a followed absolute
-	 * symlink target) restarts at "/". */
-	char abspath[MAXPATHLEN];
-	abspath[0] = '\0';
-	if (am_daemon && module_dir && module_dir[0] == '/')
-		strlcpy(abspath, module_dir, sizeof abspath);	/* "/" for a path=/ module */
-	else if (confine_root) {
-		/* Unlike a daemon's, this cwd is not pinned to the root -- the receiver
-		 * chdir's into the destination -- so it has to be read, not assumed.
-		 * It must be the PHYSICAL cwd: curr_dir is the lexical name change_dir()
-		 * was given, so after descending a trusted symlink the tracker sits at a
-		 * different depth than the kernel, and a ".." that really escapes looks
-		 * like it landed inside.
-		 *
-		 * Without it there is nothing to measure against, and an empty tracker
-		 * does NOT deny by itself -- a leading ".." pops nothing and an empty
-		 * path reads as an ancestor of the root -- so refuse the open instead. */
-		if (!getcwd(abspath, sizeof abspath))
-			return -1;
-	}
-
-	/* An fd pin (rrsync rewrites an option path to /proc/self/fd/N so no
-	 * later symlink can redirect it) is spelled outside the root by
-	 * construction, so the walk has to be allowed through /proc/self/fd to
-	 * reach the magic link.  This only suspends the check for that prefix:
-	 * following the link restarts the walk at its absolute target, and every
-	 * component of THAT is checked, so a pin aimed outside is still refused. */
-	int pin_transit = !am_daemon && confine_root && fd_pin_tail(path) != NULL;
-
-	/* Path-walk state. `remaining` is the unconsumed tail; we splice
-	 * symlink targets back into it as we go. Sized 2x MAXPATHLEN so a
-	 * one-level expansion can't immediately overflow; deeper chains
-	 * fail with ENAMETOOLONG below. */
-	char remaining[MAXPATHLEN * 2];
-	if (strlcpy(remaining, path, sizeof remaining) >= sizeof remaining) {
-		errno = ENAMETOOLONG;
-		return -1;
-	}
-
-	/* Absolute path: pin "/" as the starting dfd. */
-	if (remaining[0] == '/') {
-		dfd = open("/", O_RDONLY | O_DIRECTORY | O_CLOEXEC);
-		if (dfd < 0)
-			return -1;
-		dfd_owns = 1;
-		abspath[0] = '\0';			/* now resolving from "/" */
-		char *p = remaining;
-		while (*p == '/') p++;
-		memmove(remaining, p, strlen(p) + 1);
-	}
-
-	int loops = 40;	/* SYMLOOP_MAX-ish; breaks symlink cycles. Counts symlink
-			 * expansions only (below), NOT path depth -- a deep but
-			 * symlink-free path must resolve, not ELOOP. */
-	int retfd = -1;
-	int saved_errno = 0;
-
-	while (*remaining) {
-		/* Peel one component off the front of `remaining`. */
-		char *slash = strchr(remaining, '/');
-		size_t comp_len = slash ? (size_t)(slash - remaining) : strlen(remaining);
-		char comp[MAXPATHLEN];
-		if (comp_len == 0 || comp_len >= sizeof comp) {
-			saved_errno = comp_len == 0 ? EINVAL : ENAMETOOLONG;
-			goto out;
-		}
-		memcpy(comp, remaining, comp_len);
-		comp[comp_len] = '\0';
-		int is_last = (slash == NULL);
-
-		/* Inspect this component without following symlinks. */
-		STRUCT_STAT lst;
-		if (fstatat(dfd, comp, &lst, AT_SYMLINK_NOFOLLOW) < 0) {
-			/* The leaf may not exist yet (O_CREAT case). Allow it
-			 * and openat with O_NOFOLLOW so a race-planted leaf
-			 * symlink at this instant is still refused. */
-			if (is_last && errno == ENOENT && (flags & O_CREAT)) {
-				if (abspath_step(abspath, sizeof abspath, comp, comp_len) < 0) {
-					saved_errno = errno;
-					goto out;
-				}
-				if (!pin_transit && abspath_outside_confinement(abspath)) {
-					saved_errno = ELOOP;
-					goto out;
-				}
-				retfd = openat(dfd, comp, flags | O_NOFOLLOW, mode);
-				saved_errno = errno;
-				goto out;
-			}
-			saved_errno = errno;
-			goto out;
-		}
-
-		if (S_ISLNK(lst.st_mode)) {
-			/* Symlink: untrusted owner is refused; trusted owner
-			 * is followed via readlinkat + splice. */
-			if (lst.st_uid != 0 && lst.st_uid != trusted_uid) {
-				saved_errno = ELOOP;
-				goto out;
-			}
-			if (--loops < 0) {	/* cap symlink-follow chains */
-				saved_errno = ELOOP;
-				goto out;
-			}
-			char target[MAXPATHLEN];
-			ssize_t n = readlinkat(dfd, comp, target, sizeof target - 1);
-			if (n < 0) {
-				saved_errno = errno;
-				goto out;
-			}
-			target[n] = '\0';
-
-			/* Splice: new `remaining` = <target> + <tail-after-comp>.
-			 * Absolute target restarts the walk from "/". */
-			char tail[MAXPATHLEN];
-			tail[0] = '\0';
-			if (slash)
-				strlcpy(tail, slash, sizeof tail);
-
-			char rebuilt[MAXPATHLEN * 2];
-			if (snprintf(rebuilt, sizeof rebuilt, "%s%s",
-				     target, tail) >= (int)sizeof rebuilt) {
-				saved_errno = ENAMETOOLONG;
-				goto out;
-			}
-
-			if (target[0] == '/') {
-				if (dfd_owns) close(dfd);
-				dfd = open("/", O_RDONLY | O_DIRECTORY | O_CLOEXEC);
-				if (dfd < 0) {
-					saved_errno = errno;
-					dfd_owns = 0;
-					goto out;
-				}
-				dfd_owns = 1;
-				abspath[0] = '\0';	/* followed an absolute target: restart from "/" */
-				/* "self" resolves to "<pid>", still inside the pin;
-				 * the magic link itself lands elsewhere and ends the
-				 * exemption.  Never turns back on. */
-				pin_transit = pin_transit && fd_pin_tail(rebuilt) != NULL;
-				char *p = rebuilt;
-				while (*p == '/') p++;
-				strlcpy(remaining, p, sizeof remaining);
-			} else {
-				strlcpy(remaining, rebuilt, sizeof remaining);
-			}
-			continue;
-		}
-
-		/* Non-symlink. */
-		if (is_last) {
-			if (abspath_step(abspath, sizeof abspath, comp, comp_len) < 0) {
-				saved_errno = errno;
-				goto out;
-			}
-			if (!pin_transit && abspath_outside_confinement(abspath)) {
-				saved_errno = ELOOP;
-				goto out;
-			}
-			retfd = openat(dfd, comp, flags | O_NOFOLLOW, mode);
-			saved_errno = errno;
-			/* Resolved leaf dir (O_DIRECTORY): hand its path back so
-			 * owner_walk_parent can filter-check the operation's leaf. */
-			if (retfd >= 0 && out_abs && out_cap)
-				/* Root-resolved (".." popped abspath empty) tracked daemon walk:
-				 * hand back "/" so owner_walk_parent still leaf-checks (path=/ bypass). */
-				strlcpy(out_abs, (am_daemon && !abspath[0]) ? "/" : abspath, out_cap);
-			goto out;
-		}
-
-		if (!S_ISDIR(lst.st_mode)) {
-			saved_errno = ENOTDIR;
-			goto out;
-		}
-		/* track the resolved path so a target outside the module is refused */
-		if (abspath_step(abspath, sizeof abspath, comp, comp_len) < 0) {
-			saved_errno = errno;
-			goto out;
-		}
-		if (!pin_transit && abspath_outside_confinement(abspath)) {
-			saved_errno = ELOOP;
-			goto out;
-		}
-		int next = openat(dfd, comp, O_RDONLY | O_DIRECTORY | O_NOFOLLOW | O_CLOEXEC);
-		if (next < 0) {
-			saved_errno = errno;
-			goto out;
-		}
-		if (dfd_owns) close(dfd);
-		dfd = next;
-		dfd_owns = 1;
-
-		/* Advance `remaining` past this component (and the slash). */
-		if (slash) {
-			char *p = slash;
-			while (*p == '/') p++;
-			memmove(remaining, p, strlen(p) + 1);
-		} else {
-			remaining[0] = '\0';
-		}
-	}
-
-	/* Path resolved entirely to a directory (no leaf component left).
-	 * If the caller wanted O_DIRECTORY we already hold the dirfd we
-	 * built up; otherwise it's an EISDIR. */
-	if (flags & O_DIRECTORY) {
-		retfd = dfd;
-		dfd_owns = 0;	/* caller now owns it */
-		saved_errno = 0;
-		if (out_abs && out_cap)
-			/* Root-resolved (".." popped abspath empty) tracked daemon walk:
-			 * hand back "/" so owner_walk_parent still leaf-checks (path=/ bypass). */
-			strlcpy(out_abs, (am_daemon && !abspath[0]) ? "/" : abspath, out_cap);
-	} else {
-		saved_errno = EISDIR;
-	}
-
-out:
-	if (dfd_owns) close(dfd);
-	errno = saved_errno;
-	return retfd;
-#else
-	/* Pre-AT_FDCWD / no O_NOFOLLOW systems: best-effort fallback. */
-	(void)out_abs; (void)out_cap;
-	return open(path, flags, mode);
-#endif
-}
-
-int open_no_attacker_symlinks(const char *path, int flags, mode_t mode)
-{
-	return ona_open(path, flags, mode, NULL, 0);
-}
-
-/* When set, the do_*_at() wrappers resolve their path as an OPERATOR-supplied
- * directory path (an absolute or relative --backup-dir/--temp-dir/--*-dest)
- * using the ownership walk -- follow a symlink owned by uid 0 or our euid,
- * refuse any other-uid one, at every component -- instead of the stricter
- * transfer-path resolver (which refuses all symlinks and is confined beneath the
- * transfer root).  An operator path may legitimately point outside the tree, so
- * the trust signal is authority (ownership), not location.  Set around the
- * relevant ops by backup.c et al.; the opt-out (--insecure-links / "insecure
- * links =") restores legacy following.  Default 0 (transfer-path resolver). */
-int operator_path_resolve = 0;
-
-#if defined AT_FDCWD && defined O_NOFOLLOW && defined O_DIRECTORY
-/* For an operator-supplied path: open its parent directory via the ownership
- * walk (handles absolute and relative paths) and point *bname at the final
- * component.  Returns the dirfd (caller closes) or -1 with errno set. */
-int owner_walk_parent(const char *path, const char **bname)
-{
-	const char *slash = strrchr(path, '/');
-	char dir[MAXPATHLEN], pabs[MAXPATHLEN];
-	size_t dlen;
-	int dfd;
-
-	*bname = slash ? slash + 1 : path;
-	pabs[0] = '\0';
-	if (!slash)
-		dfd = ona_open(".", O_RDONLY | O_DIRECTORY, 0, pabs, sizeof pabs);
-	else {
-		dlen = slash == path ? 1 : (size_t)(slash - path); /* "/x" -> parent "/" */
-		if (dlen >= sizeof dir) {
-			errno = ENAMETOOLONG;
-			return -1;
-		}
-		memcpy(dir, path, dlen);
-		dir[dlen] = '\0';
-		dfd = ona_open(dir, O_RDONLY | O_DIRECTORY, 0, pabs, sizeof pabs);
-	}
-	if (dfd < 0)
-		return -1;
-	/* owner_walk only resolved the PARENT; check the resolved leaf too, so a
-	 * symlinked operator path cannot act on a leaf that resolves OUTSIDE the
-	 * module in an otherwise-served dir.  (The module exclude/filter is name-
-	 * based and not enforced here -- see abspath_outside_confinement.) */
-	if (pabs[0]) {
-		char leafabs[MAXPATHLEN];
-		if (snprintf(leafabs, sizeof leafabs, "%s/%s", pabs, *bname) >= (int)sizeof leafabs) {
-			close(dfd);
-			errno = ENAMETOOLONG;	/* fail closed, never skip the check */
-			return -1;
-		}
-		if (abspath_outside_confinement(leafabs)) {
-			close(dfd);
-			errno = ELOOP;
-			return -1;
-		}
-	}
-	return dfd;
-}
-#endif
 
 #ifndef S_BLKSIZE
 # if defined hpux || defined __hpux__ || defined __hpux
@@ -479,7 +118,6 @@ int do_unlink(const char *path)
 int do_unlink_at(const char *path)
 {
 #ifdef AT_FDCWD
-	extern int am_daemon;
 	char dirpath[MAXPATHLEN];
 	const char *bname;
 	const char *slash;
@@ -494,7 +132,7 @@ int do_unlink_at(const char *path)
 	if (operator_path_resolve) {
 		if (vfs_symlink_optout_allowed())
 			return unlink(path);
-		dfd = owner_walk_parent(path, &bname);
+		dfd = vfs_owner_walk_parent(path, &bname);
 		if (dfd < 0)
 			return -1;
 		ret = unlinkat(dfd, bname, 0);
@@ -586,7 +224,6 @@ int do_symlink(const char *lnk, const char *path)
 int do_symlink_at(const char *lnk, const char *path)
 {
 #ifdef AT_FDCWD
-	extern int am_daemon;
 	char dirpath[MAXPATHLEN];
 	const char *bname;
 	const char *slash;
@@ -758,7 +395,6 @@ int do_link(const char *old_path, const char *new_path)
 int do_link_at(const char *old_path, const char *new_path)
 {
 #if defined AT_FDCWD && defined HAVE_LINKAT
-	extern int am_daemon;
 	char old_dirpath[MAXPATHLEN], new_dirpath[MAXPATHLEN];
 	const char *old_bname, *new_bname;
 	const char *old_slash, *new_slash;
@@ -782,10 +418,10 @@ int do_link_at(const char *old_path, const char *new_path)
 	if (operator_path_resolve) {
 		if (vfs_symlink_optout_allowed())
 			return do_link(old_path, new_path);
-		old_dfd = owner_walk_parent(old_path, &old_bname);
+		old_dfd = vfs_owner_walk_parent(old_path, &old_bname);
 		if (old_dfd < 0)
 			return -1;
-		new_dfd = owner_walk_parent(new_path, &new_bname);
+		new_dfd = vfs_owner_walk_parent(new_path, &new_bname);
 		if (new_dfd < 0) {
 			e = errno;
 			close(old_dfd);
@@ -927,7 +563,6 @@ int do_lchown(const char *path, uid_t owner, gid_t group)
 int do_lchown_at(const char *fname, uid_t owner, gid_t group)
 {
 #ifdef AT_FDCWD
-	extern int am_daemon;
 	char dirpath[MAXPATHLEN];
 	const char *bname;
 	const char *slash;
@@ -1057,7 +692,6 @@ int do_mknod_at(const char *pathname, mode_t mode, dev_t dev)
 	/* HAVE_MKNODAT: older Darwin declares AT_FDCWD but not mknodat(), so
 	 * the at-variant won't build there; fall back to do_mknod() (#896). */
 #if defined(AT_FDCWD) && defined(HAVE_MKNODAT)
-	extern int am_daemon;
 	char dirpath[MAXPATHLEN];
 	const char *bname;
 	const char *slash;
@@ -1072,7 +706,7 @@ int do_mknod_at(const char *pathname, mode_t mode, dev_t dev)
 	if (operator_path_resolve) {
 		if (vfs_symlink_optout_allowed())
 			return do_mknod(pathname, mode, dev);
-		dfd = owner_walk_parent(pathname, &bname);
+		dfd = vfs_owner_walk_parent(pathname, &bname);
 		if (dfd < 0)
 			return -1;
 		if (am_root < 0) {
@@ -1204,7 +838,6 @@ int do_rmdir(const char *pathname)
 int do_rmdir_at(const char *pathname)
 {
 #ifdef AT_FDCWD
-	extern int am_daemon;
 	char dirpath[MAXPATHLEN];
 	const char *bname;
 	const char *slash;
@@ -1299,7 +932,6 @@ int do_open(const char *pathname, int flags, mode_t mode)
 int do_open_at(const char *pathname, int flags, mode_t mode)
 {
 #ifdef AT_FDCWD
-	extern int am_daemon;
 	char dirpath[MAXPATHLEN];
 	const char *bname;
 	const char *slash;
@@ -1315,7 +947,7 @@ int do_open_at(const char *pathname, int flags, mode_t mode)
 	if (operator_path_resolve) {
 		if (vfs_symlink_optout_allowed())
 			return do_open(pathname, flags, mode);
-		dfd = owner_walk_parent(pathname, &bname);
+		dfd = vfs_owner_walk_parent(pathname, &bname);
 		if (dfd < 0)
 			return -1;
 		ret = openat(dfd, bname, flags | O_NOFOLLOW, mode);
@@ -1572,7 +1204,6 @@ static int do_fchmodat_nofollow(int dfd, const char *name, mode_t mode)
 int do_chmod_at(const char *fname, mode_t mode)
 {
 #ifdef AT_FDCWD
-	extern int am_daemon;
 	char dirpath[MAXPATHLEN];
 	const char *bname;
 	const char *slash;
@@ -1668,7 +1299,6 @@ int do_rename(const char *old_path, const char *new_path)
 int do_rename_at(const char *old_path, const char *new_path)
 {
 #ifdef AT_FDCWD
-	extern int am_daemon;
 	char old_dirpath[MAXPATHLEN], new_dirpath[MAXPATHLEN];
 	const char *old_bname, *new_bname;
 	const char *old_slash, *new_slash;
@@ -1693,10 +1323,10 @@ int do_rename_at(const char *old_path, const char *new_path)
 	if (operator_path_resolve) {
 		if (vfs_symlink_optout_allowed())
 			return do_rename(old_path, new_path);
-		old_dfd = owner_walk_parent(old_path, &old_bname);
+		old_dfd = vfs_owner_walk_parent(old_path, &old_bname);
 		if (old_dfd < 0)
 			return -1;
-		new_dfd = owner_walk_parent(new_path, &new_bname);
+		new_dfd = vfs_owner_walk_parent(new_path, &new_bname);
 		if (new_dfd < 0) {
 			e = errno;
 			close(old_dfd);
@@ -1868,7 +1498,6 @@ int do_mkdir(char *path, mode_t mode)
 int do_mkdir_at(char *path, mode_t mode)
 {
 #ifdef AT_FDCWD
-	extern int am_daemon;
 	char dirpath[MAXPATHLEN];
 	const char *bname;
 	const char *slash;
@@ -1884,7 +1513,7 @@ int do_mkdir_at(char *path, mode_t mode)
 	if (operator_path_resolve) {
 		if (vfs_symlink_optout_allowed())
 			return mkdir(path, mode);
-		dfd = owner_walk_parent(path, &bname);
+		dfd = vfs_owner_walk_parent(path, &bname);
 		if (dfd < 0)
 			return -1;
 		ret = mkdirat(dfd, bname, mode);
@@ -1999,7 +1628,6 @@ int do_lstat(const char *path, STRUCT_STAT *st)
 static int do_xstat_at(const char *path, STRUCT_STAT *st, int at_flags, int (*fallback)(const char *, STRUCT_STAT *))
 {
 #ifdef AT_FDCWD
-	extern int am_daemon;
 	char dirpath[MAXPATHLEN];
 	const char *bname;
 	const char *slash;
@@ -2010,7 +1638,7 @@ static int do_xstat_at(const char *path, STRUCT_STAT *st, int at_flags, int (*fa
 	if (operator_path_resolve) {
 		if (vfs_symlink_optout_allowed())
 			return fallback(path, st);
-		dfd = owner_walk_parent(path, &bname);
+		dfd = vfs_owner_walk_parent(path, &bname);
 		if (dfd < 0)
 			return -1;
 		ret = fstatat(dfd, bname, st, at_flags);
@@ -2246,7 +1874,6 @@ int do_utimensat(const char *path, STRUCT_STAT *stp)
 int do_utimensat_at(const char *path, STRUCT_STAT *stp)
 {
 #ifdef AT_FDCWD
-	extern int am_daemon;
 	struct timespec t[2];
 	char dirpath[MAXPATHLEN];
 	const char *bname;
@@ -2660,7 +2287,7 @@ int secure_mkstemp(char *template, mode_t perms, int operator_path)
 			dir = dirbuf;
 		}
 		dirfd = operator_path
-		      ? open_no_attacker_symlinks(dir, O_RDONLY | O_DIRECTORY, 0)
+		      ? vfs_open_owner_walk(dir, O_RDONLY | O_DIRECTORY, 0)
 		      : vfs_resolve_open(dir, ".", O_RDONLY | O_DIRECTORY, 0);
 		if (dirfd < 0)
 			return -1;
@@ -2717,7 +2344,6 @@ int do_open_checklinks(const char *pathname)
 int open_dir_secure(const char *dirname)
 {
 #ifdef AT_FDCWD
-	extern int am_daemon;
 	int dfd;
 
 	/* Authority gate, identical to the do_*_at() wrappers.  When hardened
